refactor: questions module refactor
