+++ src/questions/questions.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { QuestionsController } from './controller/questions/questions.controller';
-import { QuestionsService } from './services/questions/questions.service';
+import { QuestionsService } from './services/questions.service';
 import { SharedServicesModule } from 'src/shared/shared-services.module';
 
 @Module({
